+++ reinstall.sh
@@ -1,2 +1,0 @@
-pip3 uninstall sparrowcloud
-pip3 install git+https://github.com/hanguangbaihuo/sparrow_cloud.git